% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.r
\name{visualize_categorical}
\alias{visualize_categorical}
\title{Function to visualize multiple categorical variables using bar plots}
\usage{
visualize_categorical(df, vars)
}
\arguments{
\item{df}{A data frame containing the data for visualization.}

\item{vars}{A character vector of column names for visualization.}
}
\value{
A grid of ggplot2 bar plots that is displayed in the viewer.
}
\description{
This function creates a grid of bar plots for up to 9 categorical variables from a given dataframe.
It checks that all variables exist and are factor or character before plotting.
}
\examples{
visualize_categorical(as.data.frame(Titanic), c("Class", "Sex", "Age"))

}
