% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.r
\name{visualize_numerical}
\alias{visualize_numerical}
\title{Function to visualize multiple numerical variables using histograms}
\usage{
visualize_numerical(df, vars)
}
\arguments{
\item{df}{A data frame containing the data for visualization.}

\item{vars}{A character vector of column names for visualization.}
}
\value{
A grid of ggplot2 histogram plots that is displayed in the viewer.
}
\description{
This function creates a grid of histogram plots for up to 9 numeric variables from a given dataframe.
It checks that all variables exist and are numeric before plotting.
}
\examples{
visualize_numerical(mtcars, c("mpg", "hp", "wt"))

}
